% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbidostat.R
\name{combine_turbidostat}
\alias{combine_turbidostat}
\title{Combine turbidostat data with normal cultivation data}
\usage{
combine_turbidostat(df.od, df.tb, method = "id",
  keys = c("experiment_id", "channel"), ...)
}
\arguments{
\item{method}{Method of combining}

\item{...}{extra arguments passes to other functions}

\item{od.data}{A data.frame with od data}

\item{tb.data}{A data.frame with turbidostat data}
}
\description{
Combine turbidostat data with normal cultivation data
}
