% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbidostat.R
\name{combine_turbidostat_time}
\alias{combine_turbidostat_time}
\title{Merge by time}
\usage{
combine_turbidostat_time(df.od, df.tb, keys = c("experiment_id",
  "channel"), interval = 5/60, sync_time = F)
}
\description{
Merge by time
}
