% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbidostat.R
\name{count_occurence}
\alias{count_occurence}
\title{Count occurences of a value in a vector}
\usage{
count_occurence(values, value = "True", include = TRUE)
}
\arguments{
\item{values}{a vector with all the values}

\item{value}{the value to count}

\item{include}{whether to include the occurence when counting}
}
\description{
Count occurences of a value in a vector
}
