% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbidostat.R
\name{count_occurence_in}
\alias{count_occurence_in}
\title{Count the occurence of a value in a column of a data.frame}
\usage{
count_occurence_in(data, column, value = 1, include = T,
  .to = "count")
}
\arguments{
\item{data}{data.frame with the data to count occurences in}

\item{column}{name of the column to count occurences in}

\item{value}{the value to count, i.e. if this value is seen add 1 to count}

\item{include}{whether to increase the count at that position or 1 position later}

\item{.to}{name of the column to store the count value in}
}
\description{
Count the occurence of a value in a column of a data.frame
}
