% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbidostat_growth.R
\name{decision_growth}
\alias{decision_growth}
\title{Estimates growth in one dilution segment of turbidostat data.}
\usage{
decision_growth(data, x, y, method = "lsq", .f = NULL,
  .reduce = TRUE, ...)
}
\arguments{
\item{data}{data.frame with the dilution segment}

\item{x}{Name of column with time (in hours) information}

\item{y}{Name of column with biomass information}

\item{method}{Name of the method to use for growth calculation
('lm' for linear models, 'lsq' for fast least-squares, 'ts' for theil-sen)}

\item{.f}{Custom function to use growth rate estimation (overrides the method option)}

\item{.reduce}{Reduces the data to points between the lowest and highest value (by index)}

\item{...}{extra arguments passed to other functions}
}
\description{
Estimates growth in one dilution segment of turbidostat data.
}
