% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{extract_file_info}
\alias{extract_file_info}
\title{Extract file information from a file name string}
\usage{
extract_file_info(str, prefix = "latest")
}
\arguments{
\item{str}{The filename}

\item{prefix}{The prefix used to generate data file name}
}
\value{
data.frame with the information extracted from the given path
}
\description{
Extract file information from a file name string
}
