% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{extract_project_id}
\alias{extract_project_id}
\title{Extract the project_id from a filename string}
\usage{
extract_project_id(str, prefix = "latest")
}
\arguments{
\item{str}{The filename}

\item{prefix}{The prefix used to generate data file name}
}
\value{
The project_id or NA if not found
}
\description{
Extract the project_id from a filename string
}
