% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{extract_t_zero}
\alias{extract_t_zero}
\title{Extract the time zero timestamp from a file name string}
\usage{
extract_t_zero(str, prefix = "latest")
}
\arguments{
\item{str}{The filename}

\item{prefix}{The prefix used to generate data file name}
}
\value{
The timestamp or NA if not found
}
\description{
Extract the time zero timestamp from a file name string
}
