% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{find_pycultivator_data}
\alias{find_pycultivator_data}
\title{Locate pycultivator compitable files in the given directory}
\usage{
find_pycultivator_data(path = "./data", types = c(),
  prefix = "latest", ...)
}
\arguments{
\item{path}{The directory that will be searched}

\item{types}{The type of data files to search for}

\item{prefix}{The prefix used to generate data file name}
}
\value{
data.frame with the found files and annotations
}
\description{
Locate pycultivator compitable files in the given directory
}
