% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{find_pycultivator_sqlite}
\alias{find_pycultivator_sqlite}
\title{Locate pycultivator compatible SQLITE files in the given directory}
\usage{
find_pycultivator_sqlite(path = "./data", prefix = "latest",
  rm.na = TRUE)
}
\arguments{
\item{path}{The directory with the files}

\item{prefix}{Prefix used to generate data file name}

\item{rm.na}{logical indicatng whether to remove data-files whose file name could not parsed (default=TRUE)}
}
\value{
data.frame with the found files and annotations
}
\description{
Locate pycultivator compatible SQLITE files in the given directory
}
