% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth.lm}
\alias{growth.lm}
\title{Estimates Exponential Growth using a Linear Regression Model
Applies a linear model to the given data.frame}
\usage{
growth.lm(data, x, y, .transform = TRUE)
}
\arguments{
\item{data}{data.frame contain growth data}

\item{x}{Name of column containing time information (in hours)}

\item{y}{Name of column containin biomass information}

\item{.transform}{Whether to apply `log` transformation to the `y` column}
}
\value{
Linear Model
}
\description{
Estimates Exponential Growth using a Linear Regression Model
Applies a linear model to the given data.frame
}
