% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth.lsq}
\alias{growth.lsq}
\title{Estimates Exponential Growth using a Least Squares Regression Model
Applies the least squares regression model to the given data.frame
NOTE: this is stripped (faster) version of regular Linear Models}
\usage{
growth.lsq(data, x, y, .transform = TRUE)
}
\arguments{
\item{data}{data.frame contain growth data}

\item{x}{Name of column containing time information (in hours)}

\item{y}{Name of column containin biomass information}

\item{.transform}{Whether to apply `log` transformation to the `y` column}
}
\value{
data.frame with estimated coefficients and r.squared
}
\description{
Estimates Exponential Growth using a Least Squares Regression Model
Applies the least squares regression model to the given data.frame
NOTE: this is stripped (faster) version of regular Linear Models
}
