% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth.ts}
\alias{growth.ts}
\title{Estimates Exponential Growth using a Theil-Sen Regression Model
Applies the theil-sen regression model to the given data.frame}
\usage{
growth.ts(data, x, y, .transform = TRUE)
}
\arguments{
\item{data}{data.frame contain growth data}

\item{x}{Name of column containing time information (in hours)}

\item{y}{Name of column containin biomass information}

\item{.transform}{Whether to apply `log` transformation to the `y` column}
}
\value{
Theil-Sen model
}
\description{
Estimates Exponential Growth using a Theil-Sen Regression Model
Applies the theil-sen regression model to the given data.frame
}
