% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{lm.lsq}
\alias{lm.lsq}
\title{Calculates the first degree least-squares regression model}
\usage{
lm.lsq(formula, data)
}
\arguments{
\item{formula}{formula defining which columns to use from the given data frame}

\item{data}{data.frame containing the columns specified in the formula}
}
\description{
Calculates the first degree least-squares regression model
}
