% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_pycultivator_csv}
\alias{load_pycultivator_csv}
\title{Load cultivation data from a, with pyCultivator generated, CSV file}
\usage{
load_pycultivator_csv(path, t_zero = NA,
  time_fmt = "\%Y-\%m-\%d \%H:\%M:\%S", sep = ";", dec = ",")
}
\arguments{
\item{path}{A string with the path to the data file}

\item{t_zero}{A string representing the time zero timepoint or
NA to use the first timepoint in the dataset}

\item{time_fmt}{Format to use when parsing Date/Time strings}
}
\description{
Load cultivation data from a, with pyCultivator generated, CSV file
}
