% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_pycultivator_data}
\alias{load_pycultivator_data}
\title{Load cultivation data from a, with pyCultivator generated, data source}
\usage{
load_pycultivator_data(path, t_zero = NA,
  time_fmt = "\%Y-\%m-\%d \%H:\%M:\%S", type = "csv")
}
\arguments{
\item{path}{A string with the path to the data file}

\item{t_zero}{A string representing the time zero timepoint or
NA to use the first timepoint in the dataset}

\item{time_fmt}{Format to use when parsing Date/Time strings}

\item{type}{File format to use when reading the file
('sqlite' for sqlite and 'csv' for csv files.)}
}
\value{
data.frame The retrieved data
}
\description{
Load cultivation data from a, with pyCultivator generated, data source
}
