% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbidostat.R
\name{load_turbidostat_sqlite}
\alias{load_turbidostat_sqlite}
\title{Load turbidostat data from a, with pyCultivator generated, SQLITE file}
\usage{
load_turbidostat_sqlite(path, t_zero = NA,
  time_fmt = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{path}{A string with the path to the data file}

\item{t_zero}{A string representing the zero timepoint or NA
when the first timepoint in the dataset should be taken as zero}

\item{time_fmt}{Format to use when parsing Date/Time string}
}
\description{
Load turbidostat data from a, with pyCultivator generated, SQLITE file
}
