% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{mmp_project_template}
\alias{mmp_project_template}
\title{Project Template function}
\usage{
mmp_project_template(name, folder = getwd(), readme = TRUE,
  git = TRUE, packrat = TRUE)
}
\description{
Project Template function
}
