% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmpr.R
\docType{package}
\name{mmpr}
\alias{mmpr}
\alias{mmpr-package}
\title{mmpr: A package for analyzing data obtained from the pyCultivator}
\description{
mmpr: A package for analyzing data obtained from the pyCultivator
}
