% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{parse_pycultivator_csv}
\alias{parse_pycultivator_csv}
\title{Parse data obtained from a pycultviator CSV file}
\usage{
parse_pycultivator_csv(data, t_zero = NA,
  time_fmt = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\description{
Parse data obtained from a pycultviator CSV file
}
