% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{parse_pycultivator_sqlite}
\alias{parse_pycultivator_sqlite}
\title{Parse data obtained from a, with pyCultivator generated, SQLITE file}
\usage{
parse_pycultivator_sqlite(data, t_zero = NA,
  time_fmt = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\description{
Parse data obtained from a, with pyCultivator generated, SQLITE file
}
