% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbidostat.R
\name{parse_turbidostat}
\alias{parse_turbidostat}
\title{Parse data obtained from a turbidostat SQLite file,
  generated by pyCultivator}
\usage{
parse_turbidostat(data, t_zero = NA,
  time_fmt = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{data}{data.frame with turbidostat data to parse.}

\item{t_zero}{timestamp of experiment startpoint
if NA, determine from data}

\item{time_fmt}{character format specifiction of t_zero.}
}
\description{
Parse data obtained from a turbidostat SQLite file,
  generated by pyCultivator
}
