% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{simulate_growth}
\alias{simulate_growth}
\title{Simulate exponential growth}
\usage{
simulate_growth(t, mu, initial = 1, .error.range = 0, .error.sd = 1)
}
\arguments{
\item{t}{A vector with time points (in hours)}

\item{mu}{Growth rate to grow at}

\item{initial}{Initial amount of biomass}

\item{error.range}{Range of error values}

\item{error.sd}{Standard deviation of error values}
}
\description{
Can provide noise to the data using `.error.range`
  and `.error.sd`. Uses rnorm to add random noise.
}
