% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_growth.R
\name{slide_growth_window}
\alias{slide_growth_window}
\title{Estimate exponential growth rate using a sliding window.
The growth rate is estimates for every window generated on the data.}
\usage{
slide_growth_window(data, x, y, w_size = 10, align = "center",
  method = "lsq", .to = "model", .tidy = TRUE, .transform = TRUE,
  ...)
}
\arguments{
\item{data}{data.frame with all growth data.}

\item{x}{char Name of column in data with time information.}

\item{y}{char Name of column in data with biomass information.}

\item{w_size}{Size of the sliding window}

\item{align}{Alignment of the sliding window, relative to key time.}

\item{method}{char Method used to determine growth rate in each window.}

\item{.to}{Name of column to store the model in}

\item{.tidy}{Whether to expand the models into columns.}

\item{.transform}{Whether to apply log transformation to the `y` column.}

\item{...}{extra arguments passed to slide_windows}
}
\value{
Nested data.frame with a linear model for each window.
}
\description{
Estimate exponential growth rate using a sliding window.
The growth rate is estimates for every window generated on the data.
}
