% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_growth.R
\name{slide_growth_windows}
\alias{slide_growth_windows}
\title{Estimate exponential growth rate using multiple sliding windows}
\usage{
slide_growth_windows(data, x, y, w_sizes = NULL, w_min = 10,
  w_max = NULL, w_step = 10, align = "center", method = "lsq",
  .to = "model", .tidy = TRUE, .transform = TRUE, ...)
}
\arguments{
\item{data}{data.frame with all growth data.}

\item{x}{char Name of column in data with time information.}

\item{y}{char Name of column in data with biomass information.}

\item{w_sizes}{integer vector with all the desired window sizes.}

\item{w_min}{smallest window size}

\item{w_max}{largest window size}

\item{w_step}{window size increase}

\item{align}{Alignment of the sliding window, relative to key time.}

\item{method}{char Method used to determine growth rate in each window.}

\item{.to}{Name of column to store the model in}

\item{.tidy}{Whether to expand the models into columns.}

\item{.transform}{Whether to apply log transformation to the `y` column.}
}
\description{
Window sizes can set a-priori as a vector of desired sizes,
  using the `w_sizes` argument. It can systematically
  vary window sizes from a minimum to maximum in steps.
  Use `w_min`, `w_max` and `w_step` for this.
}
