% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{stamp_filename}
\alias{stamp_filename}
\title{Make timestamped filename}
\usage{
stamp_filename(name, stamp = NULL, .stamp.format = "\%Y\%m\%d",
  .path = NULL)
}
\arguments{
\item{name}{character name of the file (with extension)}

\item{stamp}{object which can be converted to `POSIXlt`}

\item{.stamp.format}{character string specifying the format of the stamp. Defaults to `\%Y\%m\%d`.}

\item{.path}{character string specifying an addition path that should be prepended to the name}
}
\value{
character string with the timestamped file path
}
\description{
Make timestamped filename
}
