% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_growth.R
\name{tidy.sen}
\alias{tidy.sen}
\title{Extract the Theil-Sen model statistics from a Theil-Sen model object.}
\usage{
tidy.sen(model)
}
\arguments{
\item{model}{A Theil-Sen model obtained with the 'zyp' package}
}
\description{
Extract the Theil-Sen model statistics from a Theil-Sen model object.
}
