% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_growth.R
\name{tidy_slide_growth.lsq}
\alias{tidy_slide_growth.lsq}
\title{Tidy nested Least-Squares Regression Models.}
\usage{
tidy_slide_growth.lsq(model, x, ...)
}
\arguments{
\item{model}{The model to tidy}

\item{x}{reference to the column used to create windows}

\item{...}{extra arguments passes to other functions}
}
\description{
Tidy nested Least-Squares Regression Models.
}
