% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_growth.R
\name{tidy_slide_growth}
\alias{tidy_slide_growth}
\title{Tidy the nested growth rate models obtained from sliding windows}
\usage{
tidy_slide_growth(model, x, method = "lsq", .f = NULL, ...)
}
\arguments{
\item{model}{data.frame The model to tidy}

\item{x}{name of column that is the x-variable in the model}

\item{method}{name of method to tidy models with}

\item{.f}{reference to function used to tidy the models}

\item{...}{extra arguments passes to other functions}
}
\description{
Tidy the nested growth rate models obtained from sliding windows
}
