% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbidostat_growth.R
\name{turbidostat_growth}
\alias{turbidostat_growth}
\title{Calculate growth rates on turbidostat data}
\usage{
turbidostat_growth(data, x, y, method = "lsq", .group = "decision",
  .to = "model", .tidy = TRUE, .transform = TRUE, .f = NULL, ...)
}
\arguments{
\item{data}{data.frame to use}

\item{x}{Name of the column with time data (in hours)}

\item{y}{Name of the column with OD data}

\item{method}{Name of the method to use for growth calculation (see `decision_growth` for options)}

\item{.group}{Name of the column to group the data by and calculate one growth rate per group}

\item{.to}{Name of the column to store the model in}

\item{.tidy}{Whether to tidy the model into a data.frame}

\item{.transform}{Whether to apply log transformation to the y column}

\item{.f}{function used to calculate growth with (if NULL use decision_growth)}

\item{...}{extra arguments passes to other functions}
}
\description{
This method uses a grouping factor to
nest the data.frame (by default, decision). Then a function is applied
to each nested data.frame to estimate the growth rate for that decision.
}
\details{
By default, the `decision_growth` function is used to estimate growth rate,
but custom functions can be passed using the `.f` argument.
}
